Restructured to look like a more resolved backend product server by separating layers of responsibility.
