+++ src/auxiliary/generateProducts.js
