+++ src/auxiliary/simulateActivity.js
