+++ src/auxiliary/simulateCartActivity.js
