+++ src/persistence/mongo/managers/product.manager.js
@@ -1,7 +1,7 @@
 import { v4 as uuidv4 } from "uuid";
 import { connectDB } from "../connectors/mongo.connector.js";
 import { ProductModel } from "../models/product.model.js";
-import { notifyProductChange } from "../server.js";
+import { notifyProductChange } from "../../../server.js";
 
 function validateCompletenessOfProduct(Product){
     let result = false
